refactor(runtime): 提取 RuntimeToolRow 组件并统一 settings 页面间距
