+++ src/renderer/pages/security/index.tsx
@@ -226,7 +226,7 @@ export default function SecurityPage() {
 
   return (
     <PageWrapper title={t('settings.security')} subtitle={t('settings.securitySettings.subtitle')}>
-      <div className='flex flex-col gap-3'>
+      <div className='flex flex-col gap-3 mt-4'>
         <SecurityItem
           icon={<Shield theme='outline' size='22' />}
           title={t('settings.securitySettings.envProtection.title')}
